Move child_process wrappers to seperate file
